# CMR-4621: Make MMT tests faster