--- a/message-queue-lib/src/cmr/message_queue/test/queue_broker_wrapper.clj
+++ b/message-queue-lib/src/cmr/message_queue/test/queue_broker_wrapper.clj
@@ -102,9 +102,10 @@
          (Thread/sleep 10)
          (if (< (- (System/currentTimeMillis) start-time) ms-to-wait)
            (recur (set (current-message-states broker-wrapper)))
-           (warn (format "Waited %d ms for messages to complete, but they did not complete: %s"
+           (warn (format "Waited %d ms for messages to complete, but they did not complete. In progress: %s"
                          ms-to-wait
-                         (current-messages broker-wrapper)))))))))
+                         (pr-str (remove #(contains? terminal-states (:state %))
+                                         (current-messages broker-wrapper)))))))))))
 
 (defn- publish-to-queue
   "Publishes a message to the queue and captures the actions with the queue history."