Matthew and I found that there was an MMT test that was creating some invalid bulk update requests for CMR that resulted in a message on the CMR message queue that could not be processed. Found a bug in our local test broker that we weren't marking a request as failed after it had failed all retries. With this change the MMT tests went from:

Before:
```
Finished in 69 minutes 45 seconds (files took 4.31 seconds to load)
1550 examples, 0 failures
```
After:
```
Finished in 22 minutes 7 seconds (files took 4.08 seconds to load)
1550 examples, 0 failures
```

Now they only spend 11 seconds total in their calls to waiting for CMR indexing to complete:

```
|    11.370s |   0.025s |   0.147s |   452 | Helpers::CmrHelper#wait_for_cmr 
```