@chris-durbin 2017-11-30
> Nice catch. Is there a test we can write to mimic that scenario?
Sorry I missed this. Yes I think I can add a test. We might even want to add a validation to bulk update for the passed in concept-ids before we create a message on the queue. I'll think about this tomorrow, add a test, and/or write a ticket.